requirements.txt
# Core data
pandas==2.2.2
numpy==1.26.4

# Visualization
matplotlib==3.9.0
seaborn==0.13.2

# Machine Learning
scikit-learn==1.5.0
xgboost==2.1.1
joblib==1.4.2

# Backend
fastapi==0.115.0
uvicorn==0.30.1
pydantic==2.7.1
python-dotenv==1.0.1
requests==2.32.3

# Notebook
jupyter==1.0.0
ipykernel==6.29.4
